+++ R/methods_rstpm2.R
@@ -24,6 +24,7 @@ set_coef.aft <- set_coef.stpm2
 #' @export
 get_vcov.stpm2 <- function(model, ...) {
     insight::check_if_installed("rstpm2")
+    vcov <- sanitize_vcov(model, vcov)
     rstpm2::vcov(model)
 }
 
@@ -49,7 +50,8 @@ get_predict.stpm2 <- function(model, newdata = NULL, ...) {
     sanity_predict_numeric(pred = pred, model = model, newdata = newdata)
     out <- data.frame(
         rowid = seq_len(nrow(newdata)),
-        estimate = pred)
+        estimate = pred
+    )
     return(out)
 }