+++ R/utils.R
@@ -18,7 +18,9 @@ get_unique_index <- function(x, term_only = FALSE) {
 
     if (length(idx) == 0) {
         return(NULL)
-    } else if (length(idx) == 1) {
+    }
+
+    if (length(idx) == 1) {
         return(x[[idx]])
     }
 
@@ -30,12 +32,17 @@ get_unique_index <- function(x, term_only = FALSE) {
         }
     }
 
-    out <- apply(out, 1, paste, collapse = ", ")
+    out <- apply(out, 1, toString)
     return(out)
 }
 
 
-get_marginaleffects_attributes <- function(x, exclude = NULL, include = NULL, include_regex = NULL) {
+get_marginaleffects_attributes <- function(
+    x,
+    exclude = NULL,
+    include = NULL,
+    include_regex = NULL
+) {
     out <- list()
     attr_names <- names(attributes(x))
     attr_names <- setdiff(attr_names, exclude)
@@ -56,7 +63,7 @@ set_marginaleffects_attributes <- function(x, attr_cache, prefix = "") {
 
 
 warn_once <- function(msg, id) {
-    if (!isTRUE(getOption(id, default = TRUE))) {
+    if (!isTRUE(getOption(id, default = TRUE)) || !isTRUE(getOption("marginaleffects_safe", default = TRUE))) {
         return(invisible())
     }
     msg <- paste(msg, "This warning appears once per session.")
@@ -71,9 +78,11 @@ warn_once <- function(msg, id) {
 # Cross join a list of data tables
 # Source: https://github.com/Rdatatable/data.table/issues/1717#issuecomment-545758165
 cjdt <- function(dtlist) {
-    Reduce(function(DT1, DT2) cbind(DT1, DT2[rep(1:.N, each = nrow(DT1))]), dtlist)
+    Reduce(
+        function(DT1, DT2) cbind(DT1, DT2[rep(1:.N, each = nrow(DT1))]),
+        dtlist
+    )
 }
-
 
 
 # recurse up. mostly useful for `tinytest`
@@ -127,7 +136,10 @@ merge_by_rowid <- function(x, y) {
 is_binary <- function(x) {
     isTRUE(checkmate::check_integerish(
         x,
-        null.ok = TRUE, upper = 1, lower = 0, any.missing = FALSE
+        null.ok = TRUE,
+        upper = 1,
+        lower = 0,
+        any.missing = FALSE
     ))
 }
 
@@ -143,7 +155,10 @@ sub_named_vector <- function(x, y) {
     } else if (length(y) == length(x)) {
         return(y)
     } else {
-        stop("set_coef() substitution error. Please report on Github with a reproducible example: https://github.com/vincentarelbundock/marginaleffects/issues", call. = FALSE)
+        stop(
+            "set_coef() substitution error. Please report on Github with a reproducible example: https://github.com/vincentarelbundock/marginaleffects/issues",
+            call. = FALSE
+        )
     }
 
     return(x)
@@ -161,11 +176,13 @@ group_to_factor <- function(group, model) {
 
 ...get <- function(x, ifnotfound = NULL) {
     eval(
-        quote(if (!anyNA(.m1 <- match(.x, ...names())) && !is.null(.m2 <- ...elt(.m1))) {
-            .m2
-        } else {
-            .ifnotfound
-        }),
+        quote(
+            if (!anyNA(.m1 <- match(.x, ...names())) && !is.null(.m2 <- ...elt(.m1))) {
+                .m2
+            } else {
+                .ifnotfound
+            }
+        ),
         pairlist(.x = x[1L], .ifnotfound = ifnotfound),
         parent.frame(1L)
     )
@@ -178,11 +195,38 @@ group_to_factor <- function(group, model) {
     if (all(not_found)) {
         return(list())
     }
-    stats::setNames(lapply(found[!not_found], function(z) {
-        eval(
-            quote(...elt(.z)),
-            pairlist(.z = z),
-            parent.frame(3L)
-        )
-    }), x[!not_found])
+    stats::setNames(
+        lapply(found[!not_found], function(z) {
+            eval(
+                quote(...elt(.z)),
+                pairlist(.z = z),
+                parent.frame(3L)
+            )
+        }),
+        x[!not_found]
+    )
+}
+
+stop_deprecate <- function(old, new = NULL) {
+    if (is.null(new)) {
+        msg <- sprintf("The `%s` argument is not supported.", old)
+    } else {
+        msg <- sprintf("The `%s` argument is not supported. Please use `%s` instead.", old, new)
+    }
+    stop(msg, call. = FALSE)
+}
+
+
+stop_sprintf <- function(msg, ...) {
+    msg <- sprintf(msg, ...)
+    stop(msg, call. = FALSE)
+}
+
+
+warn_sprintf <- function(msg, ...) {
+    if (!isTRUE(getOption("marginaleffects_safe", default = TRUE))) {
+        return(invisible())
+    }
+    msg <- sprintf(msg, ...)
+    warning(msg, call. = FALSE)
 }